1.2.3: EXPERIMENTAL build for windows-x86_64
* tested only with windows 10 & 7 on Virtual Machines.
  + In this Windows release the HOME dir (where user data including
    savegames are stored) is now in Windows UserProfile Folder instead of the data
    directory (eg: 'c:\users\NAME\Local Settings\privgold120').
  + Currently non-utf8 characters written to console are truncated and cause
    exception in python (which is caught), when writing on the given file (stdout/stderr).
  + the std libc chdir()/fopen()/... do not handle utf8, then i needed dirty fix to use WCHAR
    (\_wmkdir/.....) and use VSFileSystem::vs_{chdir,fopen,...} wrappers in windows version.
    That would be so good to have a POSIX windows (we can dream).
  + updates in vegastrike sources for windows/msys2 sdk and clang 13
    (system types, pointers printf formats, aldrv,
  + use SHPathMatchSpec() on windows instead of fnmatch. This is not exactly fnmatch,
    but close (misses characters set in brackets)
    Add (not used yet) code to convert a fnmatch pattern in an extended regex pattern.

* fix: winsys/SDL1 mouse workaround on some windows and linux systems (mouse was not moving)

* fix: gui/FontHeight management in cockpit when Cockpit font is not base font.

* fix: (1.03) bug: don't crash if the savegame cannot be saved for some reasons
  (eg:System does not allow name, access denied, ...)

+++ vegastrike/src/cmd/script/script_callbacks.cpp
@@ -68,7 +68,7 @@ string varToString (varInst * vi) {
     if (vi->objectname=="string") {
       return *((string *)vi->object);
     }else {
-      return XMLSupport::tostring ((long)vi->object);
+      return XMLSupport::tostring ((size_t)vi->object);
     }
   }
 }
